pool: Fix reuse of TCP port in xrootd mover
The xrootd mover uses a private TCP port for accepting the connection
from the client. As only a single connection is expected, it would close
the server socket as soon as the connection has been established. The
rational was that we would avoid responding on the port for longer than
absolutely necessary and we would free up ports in the limited port
range for other movers.

The problem we ran into is when the same client would read multiple
files from the same pool. It could happen that both movers would pick
the same port. This is as designed as the first mover releases the port
as soon as it got its connection. However the xrootd client has a clever
optimisation where it recognizes that the second redirect goes to the
same pool and the same port. In that case it reuses the existing TCP
connection for opening the second file. This however fails as the
request goes to the wrong mover.

The solution is to not close the server socket until the file has been
transferred.

Target: trunk
Target: 1.9.6
Target: 1.9.5
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1489/