scripts: fix Solaris grep compatibility
Quite often scripts use the grep command to identify when a string is contained
within another string.  GNU grep supports the -q (or "quiet") option to simply this
usage.

Solaris native grep does not support grep, so we must redirect stdout to /dev/null.

This patch fixes the locations where "grep -q" is used by replacing them with a
plain call to grep with stdout redirected to /dev/null

Patch is for: 1.9.5, 1.9.6 and trunk.

RB: http://rb.dcache.org/r/1492/
Acked-by: Gerd