Workaround for shell differences
Turns out that the result of the following depends on the shell

f() { echo $x; }
x=1 f
echo $x

On some shells x is not defined after the call to f; in others it is.

This affects our use of the ourHomeDir variable used in some
configuration files. The patch simply ensures that it is defined once
and for all in bin/dcache.

Target: trunk