cells: Use Java Properties class for import command
Rather than using a custom parser, this patch uses the Properties class
that comes with Java for loading files like dCacheSetup. Thus we get
line continuation and other nice features for free.

As the properties are not ordered, we have to use a slightly more
complicated algorithm for resolve placeholders in the values.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1501/