info: add support for testing list-items
The info service unit testing relies on a test/trivial implementation of the StateExhibitor interface: the 
TestStateExhibitor class.  Classes that implement StateExhibitor allow a StateVisitor class to visit the 
current state.  Examples of StateVisitor classes include the serialisers and those classes that extract 
information from dCache's current state.

The StateComposite (part of the infrastructure providing the "live" data) allows branches to act as 
list-items: a list item has a class string and an ID-label string in addition to the usual name; for 
example, the pool called "pool_1" would have class-label string "pool" and ID-label string "name"; this 
allowing it to be identified as a "pool" (class-label) with the identifying feature that its "name" 
(ID-label) is "pool_1" (element-name).

This patch adds support to TestStateExhibitor so one can add list items that may be visited later.  This 
is to allow better unit-testing elsewhere (a pending patch).  This patch also adds a number of tests to 
verify the new functionality of TestStateExhibitor behaves as expected; these tests use a new visitor that 
allows discovered list information to be asserted.

Target: trunk.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1510/