unit-tests: fix storage-info used in PnfsManagerTest
One of the JUnit tests (org.dcache.tests.namespace.PnfsManagerTest) used location URIs that did not
specify an HSM instance.  This worked because the URIs were not strongly validated: it "worked by
accident".

A recent change (RB #1507, committed as r.13494 for Trunk, but was for all supported branches) introduced
more strict controls on what is accepted as a valid location URI.  With the addition of the more strict
checks, the code above started to fail.

This patch fixes the tests by introducing a String constant that is used as a prefix for building valid
storage-info URIs.

Patch is for all branches that RB #1507 was committed: 1.9.5, 1.9.6 and Trunk.

RB: http://rb.dcache.org/r/1518/
Acked-by: Tigran