info: add unit-tests for pretty-print serialiser
The PrettyPrintTextSerialiser is a StateSerialiser that creates an "ASCII art" representation of the 
dCache state (or subtree thereof).  It is accessible via the admin interface.

This patch adds units tests to capture expected behaviour.  The motivation is to capture correct behaviour 
prior to rewriting the class.

Note: the tests are fragile since they impose an ordering that is not (strictly) guaranteed.  The actual 
ordering will depend on the ordering of the HashMap iterator.  However, in practice this doesn't matter 
as, for the current implementations of Java API, the HashMap iterator gives (deterministically) the 
behaviour described in this patch.

This compromise is used because the "correct" solution (parsing the Pretty-Print output) would introduce 
too much complexity into the unit-tests, so negating much of the utility.

Patch is for trunk.

RB: http://rb.dcache.org/r/1519/
Acked-by: Gerd