info: ignore race-condition when shutting down.
The StateMaintainer class is basically a single-threaded executor that processes update requests and 
expunges metrics that have timed-out.

The entry-point for requesting some metric is updated is the enqueueUpdate method.  This places a new Task 
on the Executor that, when run, will result in the state being updated.  The StateMaintainer figures out 
when metric expulsion should take place and schedules the Task without requiring an external agent calling 
enqueueUpdate method.

If some update is underway that is (or will) result in a change when next to expunge metrics, and the 
StateMaintainer being shut-down then there is a race between the Executor thread scheduling the new 
expunge Task before the StateMaintainer is shut-down.  When the Executor thread looses the race then a 
RejectedExecutionException is thrown (and logged).

Under normal circumstances, the race is relatively unlikely.  When it does occur, it's harmless: shutting 
down StateMaintainer only happens as part of dCache shutdown, so the inability to (re-)schedule the metric 
expulsion doesn't matter.

However, the rapid start/stop behaviour of the unit-tests triggers this problem disproportionally often.  
The unexpected Exception results in the test failing, leading to unreliable tests.

The fix is to ignore the exception: the StateMaintainer is going to be shut-down anyway and, once 
shutdown, it cannot be started again.

Branches: Trunk.

RB: http://rb.dcache.org/r/1522/
Acked-by: Tigran