info: instrument the time spent visiting dCache state
Access to dCache's state is achieved using the VISITOR pattern.  In this pattern an object wishing to 
derive some information about dCache's state implements the StateVisitor interface and is passed to some 
StateExhibitor using this interface's visit method.  How this visiting is achieved is the responsibility 
of the StateExhibitor, but the State class (which implements StateExhibitor) uses locks to prevent 
concurrent access and modification.

The current State implement provides limited information about this process: there's no information about 
how long these visits take, or how long it takes to acquire the lock.

This patch creates log entries when certain events happen: when the visitor appears, when the lock is 
acquired (so visiting starts), when visiting completes and when the lock is released.  For convenience, 
the time taken to acquire the lock and the time taken for the actual visit are displayed.

Target: Trunk

RB: http://rb.dcache.org/r/1529/
Acked-by: Timur