infoprovider: add missing GlueChunkKey attributes for nearline and offline accounting
Due to limitations in the LDAP querying language, the LDAP binding of GLUE v1.3 includes the concept of a 
GlueChunk.  This is a link between a child object and the parent object in the hierarchy; it allows the 
construction of queries like "which objects exist that are children of X and have properties Y?"

The GlueSA LDAP objects should include a GlueChunkKey pointing to the parent GlueSE LDAP object.

The current info-provider publishes up to four types of GlueSA: SRM reservations, physical storage (the 
pools), offline storage and nearline storage.  The GlueSA objects representing SRM reservations and the 
pools correctly contain the GlueChunkKey but those representing nearline and offline storage did not.

This patch introduces the missing GlueChunkKey attributes.

The patch is for 1.9.5, 1.9.6 and Trunk.

RB: http://rb.dcache.org/r/1539/
Acked-by: Gerd