info: rewrite PrettyPrintSerialiser so it doesn't use visitCompositePreLastDescend
Classes that implement the StateVisitor interface may visit dCache's current state following the VISITOR 
pattern.  The state serialisers are a group of such classes that implement StateVisitor; the 
data-gathering activity of the secondary-information-providers are the others.

One method in StateVisitor is visitCompositePreLastDescend.  This method is called before processing the 
last item of a branch and is only "used" (i.e., results in some activity) by the PrettyPrintSerialiser; 
all other StateVisitor classes don't care about whether an item is the last one to be visited.

A future change will refactor StateVisitor to remove the visitCompositePreLastDescend method.  This cannot 
happen whilst the PrettyPrintSerialiser depends on this method.

This patch rewrites PrettyPrintSerialiser so it doesn't use visitCompositePreLastDescend any longer.

Patch is for trunk

RB: http://rb.dcache.org/r/1525/
Acked-by: Gerd