info-provider: publish all reservations
The info-provider publishes information about SRM space reservations.  It obtains this information from 
the info-service's XML dataset.  The XML data contains a branch for each VO with a sub-branch branch for 
an aggregated view of all reservations with the same name for that VO.  The structure is something like:

summary.reservations.by-VO
                      |
                      +-- VO [@name="dteam"]
                      |     |
                      |     +-- by-description
                      |           |
                      |           +-- reservation-group [@description="DTEAM_RAW"]
                      |           +-- reservation-group [@description="DTEAM_D1T0"]
                      |           :
                      |
                      +-- VO [@name="atlas"]
                      |     |
                      |     +-- by-description
                      |           |
                      |           +-- reservation-group [@description="ATLASDISK"]
                      :           :

There is a "VO" elements (fourth from the top in the hierarchy) for each VO and a "reservation-group" 
element for each reservation description.

The current version of Xylophone allows an arbitrary XPath expression to allow publishing of an arbitrary 
number of LDAP objects; however, when processing the resulting node-set (from the XPath expression) it 
uses a simple iterator.  The iterator first counts the number of nodes in the resulting node-set, it then 
iterates over the node-set by constructing an XPath statement for the i-th result.  It does this by simply 
appending the suffix "[1]" to the supplied XPath expression to build the XPath expression for the first 
LDAP object, appending "[2]" for the XPath expression for the second LDAP object, and so on.

For paths where all ancestor elements are unique, this works fine; however, when iterating over all space 
reservations of a site supporting reservations for more than one VO, the "VO" element is not unique: there 
will be a VO element for each VO.  In this case, the count will include all reservations but the 
constructed XPath expressions will only work for the first VO (XML document order).  The result is only 
reservations for the first VO will be published.

This patch fixes Xylophone's behaviour to support arbitrary XPath selects.  Instead of the simplistic 
iterator, it expands an XPath expression into a list of XPath expressions, one for each node in the 
result-set.  Each of these XPath expressions uniquely describes an item in the XML document.  The iterator 
then simply iterates over this list, so ensuring that every matching item is processed as expected.

Target: Trunk

RB: http://rb.dcache.org/r/1494/
Acked-by: Gerd