gplazma: Moved at least part of the gplazma code to slf4j
Besides using the SLF4J API, this patch removes the custom cell command
to set the log level of the logger in the GPLAZMA class. It was
inconsistent with the rest of dCache, modified the appender shared by
the complete domain, and was not of much use as it only modified the log
level of a single Logger (not all loggers related to gPlazma).

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/1540/