UniversalSpringCell: Add -waitForFiles option
In particular pools present the challenge of delayed mounting of file
systems: If a file system containing pool data is not available at the
time the pool cell is started, the used to block until the setup file is
availble. This presents a challenge for zero-config deployments without
a setup file and an automatically created pool directory hierarchy.

The present patch replaces the "block until setup file is found" with a
-waitForFiles parameter in UniversalSpringCell. This is much more
flexible than waiting for the setup file, as it will also handle the
case where the data/, control/ or meta/ directories are on different
file systems.

This feature needs to be explicitly enabled by the admin on a per pool
basis: Only the admin knows if there is a risk that a file system may
not be available right away, and which mount points we should wait for.
At the same time we want the zero-config deployment to be the default,
meaning we cannot provide a non-empty default value for -waitForFiles.

Testing done: Deployed on test system and verified that pools still
start. Then I configured a pool to wait for the setup file, renamed the
setup file and restarted the pool. The pool correctly logged:

22 Feb 2010 09:50:59 (pool_0) [] Pool mode changed to disabled(fetch,store,stage,p2p-client,p2p-server)
22 Feb 2010 09:50:59 (pool_0) [] File missing: /q/pool0/setup; sleeping 30 seconds
22 Feb 2010 09:51:29 (pool_0) [] File missing: /q/pool0/setup; sleeping 30 seconds

After renaing the setup file back to the original name, the pool started
and was operational again.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1548/