pool: Auto create directories and use default values from batch file
To support a zero-config deployment of pools, this patch let's the pool
create the data/, control/ and meta/ directories if they do not already
exist. The base directory is not automatically created. It has to be
created by the admin before starting the pool.

The defaults we have in setup.temp have been embedded into pool.cell.
Thus if no setup file is present, we will apply the same defaults as we
would have used had 'dcache pool create' been used. The only exception
is the pool size, which has no default. The values in the pool cell are
hard-coded for now. With the new configuration file layout I suggest we
but the values into the defaults/ directory.

Testing done: I removed a setup file from a test pool and deployed the
new version. The pool started and was functional.

Target: trunk
Patch: http://rb.dcache.org/r/1553/
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>