Pool: Make pool size less static
Pool size was previously defined through the "set max diskspace" command
in the setup file. This meant that the pool setup file was mandatory, as
it was the only way to set the pool size. Since we want to move to a
zero-configuration scheme for dCache, requiring the presence of a setup
file is a problem.

This patch restructures how we define the size of a pool. First of all
we move away from the concept of a pool size as a fixed value; instead
we interpret the "set max diskspace" command literally as providing a
maximum bound for the disk space usage. The command may be present in
the pool setup file, if such a file exists. This provides backwards
compatibility with existing deployments.

Besides the bound obtained through the "set max diskspace" command, the
file system size/free space provides an upper bound. Another bound is
provided in a configuration file (i.e. a .poollist file).

The three bounds are combined as follows: The pool is newer larger than
the sum of the current amount of used space and the amount of free space
on disk. If a smaller upper bound is provided with "set max diskspace",
then that bound is used. If no bound is provided with "set max
diskspace", but a smaller bound is provided in the static configuration,
then that bound is used. If neither bound is provided, then file system
will provide the only upper bound. This provides the desired
zero-configuration behaviour. The pool is never smaller than the amount
of used space. Thus a pool maybe temporarily be larger than the upper
bounds provided by the admin.

The bound stored in the setup file when "save" is always the one
provided with "set max diskspace", if any.

The pool will, as before, automatically reduce its size if the amount of
free space on disk decreases outside the control of dCache. It will
currently not automatically increase if more free space becomes
available. Such a feature is envisioned, but will require that we first
refine how we track the amount of used space. Thus the present patch
should be viewed as a step in the right direction, but it is not going
to be the last patch.

We should eventually add support for adding a gap between the pool size
and file system bound; this would provide free space for the meta data.

One consequence of this change is that we no longer need to recover from
overbooking of a pool: Instead we simply define the pool size such that
the pool is not overbooked. This simplifies pool initialisation.

As recovery from overbooking was the only recover flag still recognized,
Tigran recommended that we simply removed all recover flags.


Testing done: Installed on existing pools and verified that the old
bound was read. Tried to modify the bound with 'set max diskspace' and
save it with 'save'. Performed the usual transfer tests to verify that
the pool is still operational.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1549/