ftp, srmcp client: properly implement checksum feature negotiation, send or request checksum only if server support it
Currently we try to send or receive checksum unconditionally if client is directed to do so (sending checksum is on by default in srmcp, remoreGsiftp transfers). This causes the errors for transfers between dcache or srmcp and dpm server as reported by Gianni Pucciani and Jean-Philippe Baud.

I added a patch of jglobus, that allows the client to recognize the multiword elements in OPTS command, adds functions for discovering the CKSUM feature support, listing supported checksum algorithms, adds support for geting and sending chksum. GridfpClient that is used by srmcp and RemoteGridftp mover is modified to use these features and is simplified. 

Signed-off by: Gerd
http://rb.dcache.org/r/1547/