config-migration: remember to start transfermanagers and spacemanager
The previous config scripts would potentially start the transfermanagers and the spacemanager 
services inside the SRM domain, if they are not configured to start in their own domain.

The migration script currently assumes that transfermanagers and spacemanager are in their own 
domain.  If they are instantiated within the SRM domain then the generated layout file doesn't 
instantiate these services anywhere.

This patch fixes this by instantiating the two services in the SRM domain if they're not 
instatiated elsewhere.

Patch is for trunk only
RB: http://rb.dcache.org/r/1566
Acked-by: Gerd