Cleaner: Refactor RequestTracker
Moved the BroadcastRegistrationTask and PoolInformationBase out of
RequestTracker. This avoid instantiating those twice in ChimeraCleaner.

Moved BroadcastRegistrationTask to org.dcache.util. Turned
BroadcastRegistrationTask into a Runnable such that it can be used with
a ScheduledExecutorService.

Added a ScheduledExecutorService to ChimeraCleaner. It is currently
single threaded as it doesn't do anything besides registering with the
broadcast cell.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1568/