ChimeraCleaner: Fix responsiveness issue in HSM cleaner
The HSM cleaner performed database operations in the onSuccess
callback. This lead to responsiveness problems as the callback was
called from the message thread.

This patch moves the processing onto an ExecutorService.

The patch also replaces a dedicated thread for running the cleaner with
a periodic tasks submitted to a ScheduledExecutorService.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1572/