DCAP: Fix subject and access mask handling
When permission checking is delegated to PnfsManager, read checks have
to be explicitly requested by the door. The code for doing this was
present in the DCAP door, however when transfering by path the code
would never be executed as the door relies on an exception being
generated in the contructor of PnfsId and this exception would render
the code to set the access mask dead.

To check permissions in PnfsManager, the subject has to be specified in
the message. This is usually set by PnfsHandler, however the DCap door
would in several occasions bypass the PnfsHandler and send a message
directlry to PnfsManager. In those cases the subject would not be set,
thus bypassing permissions checks.

Target: trunk
Target: 1.9.6
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1573/