cells: Refactor message delivery and solve threading issues
Asynchronous message delivery with callbacks was handled in the thread
that called CellNucleus.addToEventQueue. This would either be the thread
of a tunnel or the thread of another cell in the same domain. If the
callback was slow or performed blocking operations, then this could
slow down the complete domain.

This patch adds two ExecutorServices to CellNucleus; one for message
delivery and one for callbacks. By default both are single threaded,
thus ensuring compatibility with current cells. Since callbacks have
their own ExecutorService, they can no longer block other cells.

Both executors can be replaced by the cell. This allows several
interesting delivery policies:

- Single threaded cells. A carefully designed cell could use the same
  single threaded executor for both callback and incoming messages. The cell
  would have to take care to never block, but a single threaded cell would
  not need to worry about race conditions from callbacks.

- Asynchronous callback delivery. CellNucleus has the asyncCallback
  property to let callbacks be run concurrently on thread pools. This is
  now implemented by using a thread pool executor without a thread limit.

- One thread per message delivery. Some of our cells tend to spawn new
  threads per message. Such cells have to be carefully designed to avoid
  race conditions and deadlocks. We can now support this usecase directly
  by setting a thread pool executor as the message delivery executor. The
  thread pool could have limits on the number of threads to avoid resource
  starvation.

- Upper limits on message queue. Some executors allow placing limits on
  the queue length or on memory usage by the queue.

We ought to commit this patch to stable branches too, as we have seen
evidence that the current callback delivery mechanism causes scalability
issues. The patch is however non-trivial, so I prefer to commit it on
trunk first and once 1.9.7 has proven stable then we backport the patch
to 1.9.5 and 1.9.6.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1581/