deployment: Source /etc/default/dcache and /etc/dcache.env
dCache relies on two environment variables to locate Java and itself:
JAVA_HOME and DCACHE_HOME. These have sensible defaults, but if one
doesn't have Java in the path or doesn't have dCache in /opt/d-cache/,
then these need to be defined. They cannot be defined in the dCache
configuration, as the values are needed to locate the dCache
configuration and to locate the JVM needed to parse the dCache
configuration.

Usually one would simply define those variables in an appropriate place
in the OS. /etc/profile.d/ is an obvious place to do this.

If for whatever reason one doesn't want to do this, this patch allows
one to define JAVA_HOME and DCACHE_HOME in either /etc/default/dcache or
/etc/dcache.env. /etc/default/ is used on several versions of Linux for
exactly this purpose: To specify a few values needed by the init script,
without the files containing the actual configuration of the service.
For those systems that don't use /etc/default/, I also source
/etc/dcache.env. I chose not to call the file dcache.conf, as this could
possibly cause it to be confused with /opt/d-cache/etc/dcache.conf (the
main configuration file of dCache).


Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1600/