PnfsId: First step to reduce memory consumption
PnfsId contains three representations of the same ID: A compact byte[]
form and two strings. Each string requires 4 times as much memory as the
byte array form. If we dropped those strings we could reduce memory
consumption from 276 bytes to 68 bytes on a 64 bit platform for short
PNFS IDs.

Unfortunately PnfsId is serializable, so we cannot just drop the two
fields without breaking compatibility. Instead we have to do this in two
steps. First we stop using _idString and _toString, but keep the fields
with the normal values. This ensures that old versions of the class can
still read the modified PnfsId. Then for 1.9.7 we drop the two strings.
At that point the 1.9.6-4 PnfsId will still be able to read the 1.9.7
version, as it doesn't actually use the two strings. Earlier versions
will not be able to work with 1.9.7.

To give a sense of how much we can safe: PIC got a pool with 1.9 million
files. Getting rid of these two strings will lead to a memory reduction
of 376 MB. That's a sizable reduction from such a trivial change.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1538/