pool: Relax requirements in PoolCostInfo
The values of 'removable' and 'precious' in the Account object are only
eventually consistent. Since the values are maintained by different
components, there may be brief periods in which they do not add up. Do
to this it can happen that an argument check in PoolCostInfo would fail.
The check is however not correct as the values are not supposed to be
consistent at all times.

This patch removes the consistency check in PoolCostInfo, thus
preventing the pool ping from failing. A real inconsistency will be
caught by the periodic pool consistency check.

Target: trunk
Target: 1.9.6
Target: 1.9.5
Ticket: #5488
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1599/