pool: Prepare for concurrent message processing
To address some of the responsiveness issues of pools, this patch
prepares the pool for concurrent processing of messages. The
infrastructure changes for such concurrent processing were recently
submitted in patch 1581. Before we can make use of multi-threaded
message delivery, we have to make the pool thread safe. That is achieved
by this patch.

I went through all messageArrived methods in the pool code and made sure
that they are thread safe.

I removed some dead code in the process. I also implemented
PoolCheckFreeSpaceMessage, which only had a dummy implementation.

I suggest we eventually backport this to 1.9.5, but not until we now it
is stable in 1.9.7.

Target: trunk
Target: 1.9.7
Ticket: #5414
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1601/