pool: Use thread pool for message delivery
This patch extends UniversalSpringCell such that custom thread pools for
message delivery and callbacks can be defined.

The pool is reconfigured to use a cached thread pool for both. A maximum
thread pool size of 10 was chosen arbitrarily. The value is
reconfigurable in the batch file, although we do not currently expose it
as a configuration parameter (can be changed should the need arise).

While implementing this, I realized that the worker thread limit on the
existing thread pool was not used as a limit but as the minimum size. I
changed this such that the parameter is actually an upper limit on the
number of threads in that thread pool.

Target: trunk
Target: 1.9.7
Ticket: #5414
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1603/