unittests: fix unit tests for PnfsId that check serialisation
A recent change to PnfsId resulted in a difference in how these objects are serialised.  
This is fine, provided old objects can be deserialised by the new code and the new objects 
can be deserialised by the old code.  We can't check the latter in unit tests, but we can 
check the former.

This patch introduces an array of serialised PnfsId objects that should be equal to the 
example PnfsId object.  It also updates the expected form of the serialisation, so the 
serialisation tests will pass.

Patch is for 1.9.7 and trunk

RB: http://rb.dcache.org/r/1608/
Acked-by: Gerd