config migration: allow from RPM
The script migrate-config-to-197.sh migrates existing (pre-1.9.7) dCache configuration to 
something that 1.9.7 will understand.  The script includes a safety feature that prevents 
someone overwriting an existing /opt/d-cache/etc/dcache.conf file.  We package a default 
dcache.conf file.  This includes some comments describing the file's purpose but is 
otherwise void of content.

The problem is that, when upgrading dCache to 1.9.7, the RPM will supply a default 
dcache.conf file.  This file will prevent the migrate-config-to-197.sh script from 
running.  The solutions to the site-admin is either to delete the RPM-supplied default 
dcache.conf file or to run the migration script with "-f" to force the migration.  Neither 
solution is desirable as both solutions are "surprising".

This patch introduces a more relaxed test for dcache.conf.  With this patch, the 
migrate-config-to-197.sh script is allow to overwrite an existing dcache.conf file if it 
contains only lines that are either empty or a comment.  This allows the 
migrate-config-to-197.sh script to be run after upgrading without any further effort, but 
prevents loosing configuration by running the script when dcache.conf has content.

Patch is for trunk (and 1.9.7, a branch created after RB review)

RB: http://rb.dcache.org/r/1570/
Acked-by: Gerd