poolcost: ignore failing unit test
For better-or-worse, the current implementation of the pool accounting infrastructure
doesn't guarantee (instantaneous) measurements are consistent.  Therefore, the
testTotalLessThanSum unit test is inappropriate.

Rather than simply deleting the test, this patch marked it "@Ignore", in the hope
that, in the future, an pool accounting infrastructure may make stronger guarantees.

Patch is for 1.9.5, 1.9.6, 1.9.7 and trunk.

RB: http://rb.dcache.org/r/1610/
Acked-by: Gerd