BootLoader: Fix quoting
The BootLoader can list configuration values in a format parseable by the
shell. For this to work certain characters in the key need to be replaced
and certain characters in the value need to be escaped.

Unfortunately the old code performed the escaping on the key and not the
value. This is fixed by this patch.

Target: trunk
Target: 1.9.7
Acked-by: Paup Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1620/