infoprovider: fix lookup of namespace implementation
The pre-1.9.7 info-provider discovered which namespace implementation (PNFS or Chimera) is being used by 
looking at the name of the domain PnfsManager was running in.

With 1.9.7, migrated pre-1.9.7 configurations place PnfsManager in the same domain (namespaceDomain) 
irrespective of which namespace implementation is being run.  This breaks the previous hack.  Moreover, 
with 1.9.7 it becomes trivially easy to move services from one domain to another, so any attempt to use 
domain names would be fragile.

This patch introduces a new method: the info-provider searches for an instance of the Chimera Cleaner 
(org.dcache.chimera.namespace.ChimeraCleaner).  If found, Chimera is reported; for example:

    GlueSEImplementationVersion: 1.9.7-1 (ns=Chimera)

If not found then it searches for an instance of the PNFS cleaner (diskCacheV111.cells.CleanerV2).  If 
found, PNFS is reported; for example:

    GlueSEImplementationVersion: 1.9.7-1 (ns=PNFS)

If not then UNDEFINEDVALUE is reported; for example:

    GlueSEImplementationVersion: 1.9.7-1 (ns=UNDEFINEDVALUE)

Patch is for 1.9.7 and Trunk.

RB: http://rb.dcache.org/r/1622/
Acked-by: Gerd