cells: Get rid of three more cell interpreters
The interpreter in CellShell is the most complete of our batch
interpreters and we should try to use it rather than using custom
interpreter loops. This patch replaces three interpreters with the one
in CellShell. The patch relies on an existing hack in CellShell in which
the shell can be attached to an external interpreter.

One side effect of the patch is that the defined setup can now contain
place holders. This will be handy when improving the config files of the
httpd cell.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1630/