webadmin: hide cells that have not responded to any xgetcellinfo requests
Previously, we introduced the gPlazma cell to the "Cell Services" web-page at the request of OSG.  
More recently we received a request (http://rt.dcache.org/Ticket/Display.html?id=5217) to hide 
the gPlazma cell if the site is configured not to run gPlazma.

When discussing this apparent conflict, the decision was that the best compromise was to publish 
the gPlazma cell only if it has already responded at least once.  In other words, we take the 
successful response to xgetcellinfo to indicate that the site has gPlazma configured as a cell.

This issue has been expanded to a request that any non-configured cell should not appear in the 
"Cell Services" page (http://rt.dcache.org/Ticket/Display.html?id=5547).

This patch implements that behaviour.  It is intended for all supported branches: 1.9.5, 1.9.6, 
1.9.7 and trunk.

RB: http://rb.dcache.org/r/1626/
Ticket: 5217
ticket: 5547
Acked-by: Gerd