info: add StateGuide Interface
The dCache state, as held in the info service, uses the VISITOR pattern to extract information.  
In this pattern a class that collects information about dCache implements the StateVisitor 
interface.  It is passed to the visit method of an object from a class that implements the 
StateExhibitor interface to build up information about dCache.

The StateVisitor interface has four visit methods: two of these are to allow a StateVisitor to 
visit only a subtree within the dCache tree.  By restricting which parts of the dCache state is 
visited, it is possible to avoid visiting information that is never going to be used by the 
visitor class.

What is unfortunate is that the pattern of which elements of dCache's state are to be visited are 
hard-coded into the StateExhibitor as the overloaded visit method.

This patch introduces the StateGuide interface.  A object of a class that implements StateGuide 
can control which parts of dCache a StateVisitor will visit.  It achieves this by implementing 
the isVisitable method.  This method describes whether a StateVisitor should be exposed to some 
branch within the dCache state: returning true will result in the StateVisitor being exposed to 
the branch elements (and potentially any sub-elements thereof), returning false will result in 
the StateVisitor not visiting that branch or any subelements of that branch.

The StateVisitor interface extends StateGuide.  This means that each StateVisitor is also a 
StateGuide, controlling which parts of the dCache state it is exposed to.  The introduction of 
StateGuide is to allow code-reuse; for example, a common pattern is to be exposed to all 
descendent elements of some branch element.  Code to support this can be factored out into a 
StateGuide and StateVisitors that wish to use this visit pattern can instantiate a suitable 
StateGuide object and delegate the isVisitable decision to that object.

Patch is for trunk.

RB: http://rb.dcache.org/r/1569
Acked-by: Gerd