deploy: Make properties obsoletable and deprecatable
This patch allows deprecated and obsolete properties to be marked as
such. Such properties are prefixed by '(deprecated)' and '(obsolete)'.
Deprecated properties output a warning when being redeclared. Obsolete
properties trigger a failure.

As this patch introduces log output before the first domain is started
and even when just querying the boot loader for configuration
parameters, this patch makes several changes to the log4j
initialization. A basic configuration is initialized very early and then
replaced by the proper log4j configuration when the SystemCell is
started. The log4j configuration file is no longer passed as a system
property. The configuraton file will be monitored and reloaded when it
changes.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1651/