SpaceManager: Propagate the file size to PoolManager and pool
The SpaceManager will often know the intended file size as this is part
of the srmPrepareToPut message. This information was however not
propagated to the PoolManager or the pool. The PoolManager needs to know
so that it can take the file size into account when selecting a write
pool. The pool needs to know as it can use it to allocate enough space
on the pool right away and it can use it to detect incomplete transfers.


I request that this patch is carefully analyzed wrt. correctness. I
don't know how the SRM and the space manager interact, so I don't know
if the sizeInBytes field of File does in fact reflect the
expectedFileSize field in the srmPrepareToPut message. I don't know
whether this field is to be considered strict or just a hint. If it is
just a hint, then failing a transfer if the file size does not match
would be wrong. Finally, I don't know what the value of sizeInBytes of
File is if the expectedFileSize field is absent. expectedFileSize is an
optional field in SRM.

If expectedFileSize is just a hint and should not cause a failure if the
final size differs, then the patch needs to be modified such that the
file size is not set in StorageInfo but only specified as a hint in the
message to PoolManager.

I will delay committing this to stable branches until I have tested the 
patch in production at NDGF.

Target: trunk
Target: 1.9.7
Target: 1.9.6
Target: 1.9.5
Ticket: #5075