info: fix support for StateGuide in StateComposite
A recent patch (RB:1569, SVN r13681: "info: add StateGuide Interface") introduced the general framework for a 
visitor to guide to which parts of the dCache state it should be exposed.  Unfortunately, the submitter forgot 
to include support in the StateComposite (a.k.a branches), resulting in visitors being exposed to all of the 
dCache state, ignoring the visitor's StateGuide.

The result of this is that all dCache state is exposed through the web interface (ignoring the user-supplied 
path in the URL).  The same problem was found when exploring dCache state in the admin interface.

This patch fixes the problem by introducing support for querying the visitor whether items should be explored.

Also included is some unit tests to verify correct behaviour, both for regular visiting and 
post-StateTransition visiting.  With current status of code these tests fail.  With this patch, the tests 
succeed.

Patch is for trunk

RB: http://rb.dcache.org/r/1668
Acked-by: Jan